Fix: MockProvider dev mode, NUMERIC casts, idempotent migrations, column names
- MockProvider with_name("stripe") for dev mode (no real Stripe keys needed)
- Auto-detect mock mode when stripe_secret_key starts with "sk_test_mock"
- Add MM_MONETIZATION_SUBSCRIPTIONS_ENABLED env override (was missing)
- Cast NUMERIC(5,4) to float8 in all PG queries returning CreatorProfile
- Fix column name mismatches in monetization handlers (perks→perks_json, active→is_active)
- Make V004-V006 migrations idempotent (IF NOT EXISTS on tables and indexes)
- Fix perks serialization: String→serde_json::Value for JSONB columns

249 tests, 29 E2E checks, all passing on LAN IP (10.0.0.105)

+++ crates/mm-db/src/monetization_db.rs
@@ -233,7 +233,7 @@ impl MonetizationDb for PgMonetizationDb {
     async fn get_creator_profile(&self, user_id: &str) -> Result<Option<CreatorProfile>, MMError> {
         sqlx::query_as::<_, CreatorProfile>(
             "SELECT id, user_id, display_name, stripe_account_id, onboarding_complete,
-                    platform_fee_pct, created_at, updated_at
+                    platform_fee_pct::float8, created_at, updated_at
              FROM mm_creator_profiles WHERE user_id = $1",
         )
         .bind(user_id)
@@ -253,7 +253,7 @@ impl MonetizationDb for PgMonetizationDb {
              VALUES ($1, $2, $3)
              ON CONFLICT (user_id) DO UPDATE SET display_name = EXCLUDED.display_name
              RETURNING id, user_id, display_name, stripe_account_id, onboarding_complete,
-                       platform_fee_pct, created_at, updated_at",
+                       platform_fee_pct::float8, created_at, updated_at",
         )
         .bind(user_id)
         .bind(display_name)
@@ -813,7 +813,7 @@ impl MonetizationDb for PgMonetizationDb {
     async fn list_creators(&self, limit: i64, offset: i64) -> Result<Vec<CreatorProfile>, MMError> {
         sqlx::query_as::<_, CreatorProfile>(
             "SELECT id, user_id, display_name, stripe_account_id, onboarding_complete,
-                    platform_fee_pct, created_at, updated_at
+                    platform_fee_pct::float8, created_at, updated_at
              FROM mm_creator_profiles
              WHERE onboarding_complete = true
              ORDER BY display_name ASC
